requirements.txt
appdirs==1.4.0
Django==1.10.5
djangorestframework==3.5.3
packaging==16.8
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
